refactor: modify architecture
This is a first step to separate orderbook from trade system

+++ src/main/java/br/com/mercadolivre/broker/orderbook/vibranium/infra/repository/AskRepositoryPostgres.java
@@ -39,7 +39,8 @@ public class AskRepositoryPostgres implements AskRepository {
 
     @Override
     public void delete(Ask ask) {
-        dao.deleteById(ask.getId());
+        if (ask.getId() != null)
+            dao.deleteById(ask.getId());
     }
 
     @Override
