+++ src/main/java/br/com/mercadolivre/broker/orderbook/vibranium/infra/repository/BidRepositoryPostgres.java
@@ -39,7 +39,8 @@ public class BidRepositoryPostgres implements BidRepository {
 
     @Override
     public void delete(Bid bid) {
-        dao.deleteById(bid.getId());
+        if (bid.getId() != null)
+            dao.deleteById(bid.getId());
     }
 
     @Override
